# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Crunchy API",
        "description": "This is a sample of Crunchy API",
        "contact": {
            "name": "SmallCorridor",
            "email": "smallcorridor@gmail.com"
        },
        "version": "1.0"
    },
    "host": "crunchy-api.dev",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/post": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Retrieves the collection of Post resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/post"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/post/{postId}/addComment": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Add comment in posting image.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    },
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Please enter the postId",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Comment that needs to be added into posting image",
                        "required": true,
                        "schema": {
                            "properties": {
                                "comment": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment has successfully added.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/comment"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/post/{postId}/comments": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Show all comments by Post's Id.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    },
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Please enter the postId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "This is the data that you search.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/comment"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "404": {
                        "description": "Post ID not found."
                    }
                }
            }
        },
        "/api/v1/post/{postId}/deleteComment/{commentId}": {
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Remove the comment from posting image.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    },
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Please enter the postId",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "Please enter the commentId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comment has been deleted."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "404": {
                        "description": "Post ID or Comment ID not found."
                    }
                }
            }
        },
        "/api/v1/post/{postId}/totalLikes": {
            "get": {
                "tags": [
                    "Like"
                ],
                "summary": "Show total likes by Post's Id.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    },
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Please enter the postId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "This is the data that you search.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/like"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "404": {
                        "description": "Post ID not found."
                    }
                }
            }
        },
        "/api/v1/post/{postId}/like": {
            "post": {
                "tags": [
                    "Like"
                ],
                "summary": "Like or Dislike a posting image.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Example = Bearer(space)'your_token'",
                        "required": true,
                        "type": "string",
                        "default": "Bearer"
                    },
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Please enter the postId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/like"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        }
    },
    "definitions": {
...
        "comment": {
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "user_id": {
                    "type": "integer",
                    "format": "int32"
                },
                "post_id": {
                    "type": "integer",
                    "format": "int32"
                },
                "comment": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
        "like": {
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "user_id": {
                    "type": "integer",
                    "format": "int32"
                },
                "post_id": {
                    "type": "integer",
                    "format": "int32"
                },
                "like": {
                    "type": "boolean"
                },
                "create_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
        "post": {
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "user_id": {
                    "type": "integer",
                    "format": "int32"
                },
                "merchant_id": {
                    "type": "integer",
                    "format": "int32"
                },
                "path": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "taste": {
                    "type": "number",
                    "format": "double"
                },
                "price": {
                    "type": "number",
                    "format": "double"
                },
                "service": {
                    "type": "number",
                    "format": "double"
                },
                "create_at": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        }
    }
}